+++ src/app/dataset-block/metadata-block/components/block-header/block-header.component.ts
@@ -7,10 +7,8 @@
 
 import { BaseComponent } from "src/app/common/components/base.component";
 import { MetadataBlockFragment } from "../../../../api/kamu.graphql.interface";
-import { ChangeDetectionStrategy, Component, inject, Input, OnInit } from "@angular/core";
-import { Observable } from "rxjs";
+import { ChangeDetectionStrategy, Component, Input } from "@angular/core";
 import { DatasetInfo } from "src/app/interface/navigation.interface";
-import { BlockService } from "../../block.service";
 
 @Component({
     selector: "app-block-header",
@@ -18,14 +16,7 @@ import { BlockService } from "../../block.service";
     styleUrls: ["./block-header.component.scss"],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class BlockHeaderComponent extends BaseComponent implements OnInit {
-    public block$: Observable<MetadataBlockFragment>;
-    @Input({ required: true }) public blockHash: string;
+export class BlockHeaderComponent extends BaseComponent {
     @Input({ required: true }) public datasetInfo: DatasetInfo;
-
-    private blockService = inject(BlockService);
-
-    public ngOnInit(): void {
-        this.block$ = this.blockService.metadataBlockChanges;
-    }
+    @Input({ required: true }) public block: MetadataBlockFragment;
 }