+++ src/app/dataset-block/metadata-block/components/yaml-view-section/yaml-view-section.component.ts
@@ -5,11 +5,9 @@
  * included in the LICENSE file.
  */
 
-import { ChangeDetectionStrategy, Component, inject, OnInit } from "@angular/core";
+import { ChangeDetectionStrategy, Component, Input } from "@angular/core";
 import { MetadataBlockFragment } from "src/app/api/kamu.graphql.interface";
-import { BlockService } from "../../block.service";
 import { SupportedEvents } from "../event-details/supported.events";
-import { Observable } from "rxjs";
 import { eventsWithYamlView } from "./yaml-view-section.types";
 
 @Component({
@@ -17,19 +15,11 @@ import { eventsWithYamlView } from "./yaml-view-section.types";
     templateUrl: "./yaml-view-section.component.html",
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class YamlViewSectionComponent implements OnInit {
-    private blockService = inject(BlockService);
-    public yamlEventText$: Observable<string>;
-
-    public ngOnInit(): void {
-        this.yamlEventText$ = this.blockService.metadataBlockAsYamlChanges;
-    }
-
-    public get currentBlock(): MetadataBlockFragment {
-        return this.blockService.currentBlock;
-    }
+export class YamlViewSectionComponent {
+    @Input({ required: true }) public blockAsYaml: string;
+    @Input({ required: true }) public block: MetadataBlockFragment;
 
     public get isEventWithYamlView(): boolean {
-        return eventsWithYamlView.includes(this.currentBlock.event.__typename as SupportedEvents);
+        return eventsWithYamlView.includes(this.block.event.__typename as SupportedEvents);
     }
 }