+++ CHANGELOG.md
@@ -4,12 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.41.0] - 2025-03-07
 ### Added
+- Added resolver `BlockMetadataResolver` for metadata block page
 - Added `DEVELOPER.md` file
 - Added license header for all *.ts files
+- Account page: added account name
 ### Changed
 - Modified `README.md` file
+- Changed the method of getting route and query parameters via input decorator for some components
 ### Fixed
 - Remove "Dashboard" link from application header
 - Disabled "keywords" feature for "demo" mode